Initial Commit (Warning!)
An exact mirror of the working setup that was used for initial testing
of a vanilla Tor configuration on 50 relays. Will need to clean up and
prep it for deployment with some optimizations and reorganizations.
Warning: It should be noted that this is the code needed to work on a
vanilla installation of Mininet VM. Also, the setups for the 50 relays
are included. They will be removed in the future (most likely).
Disclaim: All programs within this repo are property of their
respective designers. As such, this code will be removed and replaced
with a dependency instead.